Use caught exception as casue for deployment exception
Currently, the causing exception (if any, `null` otherwise) of the caught exception is set as the cause of a newly created and thrown deployment exception.

This effectively removes the causing exception, alongside valuable debugging information. When the stack-trace of the deployment exception is logged, the causing exception is missing from that stack-trace.